Remove references to ec2 task
