--- example/docker/Dockerfile.hello_world
@@ -1,8 +1,0 @@
-FROM ubuntu:16.04
-
-RUN apt-get update
-RUN apt-get install -y stress
-COPY say_hello.sh /
-
-ENTRYPOINT [ "./say_hello.sh" ]
-CMD [ "1", "2", "1024", "120" ]
