+++ example/workflows/helloworld.yml
@@ -80,60 +80,3 @@ HelloWorldActivityWorkflow:
     WorkflowFailed:
       Type: Fail
       Cause: 'Workflow failed'
-
-HelloWorldOnDemandWorkflow:
-  Comment: 'Returns Hello World'
-  StartAt: StartStatus
-  States:
-    StartStatus:
-      Type: Task
-      Resource: ${SfSnsReportLambdaFunction.Arn}
-      CumulusConfig:
-        cumulus_message:
-          input: '{$}'
-      Next: HelloWorldTask
-    HelloWorldTask:
-      ResultPath: '$.payload.task_output'
-      CumulusConfig:
-        buckets: '{$.meta.buckets}'
-        provider: '{$.meta.provider}'
-        collection: '{$.meta.collection}'
-      Type: Task
-      Resource: 'arn:aws:states:::ecs:runTask.sync'
-      TimeoutSeconds: 180
-      Parameters:
-        Cluster: ${CumulusECSCluster.Arn}
-        TaskDefinition: ${HelloWorldOnDemandTaskDefinition}
-        Overrides:
-          ContainerOverrides:
-            - Name: HelloWorldOnDemand
-              'Command.$': '$.payload.commands'
-      Retry:
-        - ErrorEquals:
-            - States.Timeout
-          MaxAttempts: 1
-        - ErrorEquals:
-          - ECS.AmazonECSException
-          IntervalSeconds: 60
-          BackoffRate: 1 # The multiplier by which the retry interval increases during each attempt (2.0 by default).
-          MaxAttempts: 20
-      Next: StopStatus
-    StopStatus:
-      Type: Task
-      Resource: ${SfSnsReportLambdaFunction.Arn}
-      CumulusConfig:
-        sfnEnd: true
-        stack: '{$.meta.stack}'
-        bucket: '{$.meta.buckets.internal.name}'
-        stateMachine: '{$.cumulus_meta.state_machine}'
-        executionName: '{$.cumulus_meta.execution_name}'
-        cumulus_message:
-          input: '{$}'
-      Catch:
-        - ErrorEquals:
-          - States.ALL
-          Next: WorkflowFailed
-      End: true
-    WorkflowFailed:
-      Type: Fail
-      Cause: 'Workflow failed'
